Make IDE compilers happy
