+++ test/e2e/azure_failuredomains.go
@@ -1,3 +1,4 @@
+//go:build e2e
 // +build e2e
 
 /*
