+++ test/e2e/azure_net_pol.go
@@ -1,3 +1,4 @@
+//go:build e2e
 // +build e2e
 
 /*
