+++ test/e2e/log.go
@@ -1,3 +1,6 @@
+//go:build e2e
+// +build e2e
+
 /*
 Copyright 2021 The Kubernetes Authors.
